+++ example/lib/util/locale/localization_keys.dart
@@ -1,0 +1,9 @@
+//============================================================//
+//THIS FILE IS AUTO GENERATED. DO NOT EDIT//
+//============================================================//
+class LocalizationKeys {
+
+  static const appTitle = 'app_title';
+
+  static const welcomeMessage = 'welcome_message';
+}